requirement.txt
appnope==0.1.0
asn1crypto==0.24.0
autopep8==1.4.3
backcall==0.1.0
backports-abc==0.5
backports.ssl-match-hostname==3.7.0.1
bleach==3.1.0
brewer2mpl==1.4.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.5
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.5
django-braces==1.13.0
django-extensions==2.1.5
entrypoints==0.3
funcsigs==1.0.2
gnureadline==6.3.8
idna==2.8
importlib-metadata==0.8
ipaddress==1.0.22
ipykernel==5.1.0
ipython==7.2.0
ipython-cypher==0.2.6
ipython-genutils==0.2.0
jedi==0.13.2
jgraph==0.2.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
mysqlclient==1.4.1
nbconvert==5.4.1
nbformat==4.4.0
neo4jrestclient==2.1.1
neobolt==1.7.4
neotime==1.7.4
networkx==2.2
nose==1.3.7
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.3
path.py==11.5.0
patsy==0.5.1
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.6.1
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
py2neo==4.2.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
requests==2.21.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.3.3
